requirements.txt
torch
torchvision
transformers
datasets
pytorch-lightning
peft
gradio
fastapi
uvicorn
lightning
scikit-learn
Pillow
tqdm

